+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/panels/GraphPanelBuilder.kt
@@ -11,7 +11,10 @@ import ru.yandex.money.tools.grafana.dsl.metrics.functions.Alias
 import ru.yandex.money.tools.grafana.dsl.panels.graph.display.drawoptions.HoverTooltip
 import ru.yandex.money.tools.grafana.dsl.panels.graph.display.seriesoverrides.SeriesOverride
 import ru.yandex.money.tools.grafana.dsl.panels.graph.display.seriesoverrides.SeriesOverrideBuilder
+import ru.yandex.money.tools.grafana.dsl.panels.repeat.Repeat
+import ru.yandex.money.tools.grafana.dsl.panels.repeat.RepeatBuilder
 import ru.yandex.money.tools.grafana.dsl.time.Duration
+import ru.yandex.money.tools.grafana.dsl.variables.Variable
 
 class GraphPanelBuilder(
     private val title: String,
@@ -58,6 +61,12 @@ class GraphPanelBuilder(
 
     var rightYAxis: YAxis? = null
 
+    var xAxis: XAxis = XAxis()
+
+    var description: String? = null
+
+    private var repeat: Repeat? = null
+
     private val seriesOverrides: MutableList<SeriesOverride> = mutableListOf()
 
     override fun properties(propertiesSetter: (JSONObject) -> Unit) {
@@ -84,35 +93,44 @@ class GraphPanelBuilder(
         return this
     }
 
+    fun repeat(variable: Variable, build: RepeatBuilder.() -> Unit) {
+        val builder = RepeatBuilder(variable)
+        builder.build()
+        repeat = builder.createRepeat()
+    }
+
     internal fun createPanel() = AdditionalPropertiesPanel(
-            GraphPanel(
-                    MetricPanel(
-                            BasePanel(
-                                    id = panelLayoutGenerator.nextId(),
-                                    title = title,
-                                    position = panelLayoutGenerator.nextPosition(bounds.first, bounds.second)
-                            ),
-                            datasource = datasource,
-                            metrics = Metrics(metrics.metrics)
-                    ),
-                    timeShift = timeShift,
-                    stack = stack,
-                    legend = legend,
-                    decimals = decimals,
-                    bars = bars,
-                    lines = lines,
-                    points = points,
-                    pointradius = pointradius,
-                    nullValue = nullValue,
-                    fill = fill,
-                    lineWidth = lineWidth,
-                    staircase = staircase,
-                    hoverTooltip = hoverTooltip,
-                    aliasColors = aliasColors,
-                    leftYAxis = leftYAxis,
-                    rightYAxis = rightYAxis,
-                    seriesOverrides = seriesOverrides
-            )
+        GraphPanel(
+            MetricPanel(
+                BasePanel(
+                    id = panelLayoutGenerator.nextId(),
+                    title = title,
+                    position = panelLayoutGenerator.nextPosition(bounds.first, bounds.second),
+                    description = description
+                ),
+                datasource = datasource,
+                metrics = Metrics(metrics.metrics)
+            ),
+            timeShift = timeShift,
+            stack = stack,
+            legend = legend,
+            decimals = decimals,
+            bars = bars,
+            lines = lines,
+            points = points,
+            pointradius = pointradius,
+            nullValue = nullValue,
+            fill = fill,
+            lineWidth = lineWidth,
+            staircase = staircase,
+            hoverTooltip = hoverTooltip,
+            aliasColors = aliasColors,
+            leftYAxis = leftYAxis,
+            rightYAxis = rightYAxis,
+            seriesOverrides = seriesOverrides,
+            xAxis = xAxis,
+            repeat = repeat
+        )
     ) { json -> propertiesSetters.forEach { it(json) } }
 }